// admin/src/api/generated/schemas/adminAdminUserItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminAdminUserItem = {
  admin_id?: string;
  username?: string;
  name?: string;
  role?: string;
  status?: string;
  two_factor_enabled?: boolean;
  last_login_at?: string;
  last_login_ip?: string;
  created_at?: string;
  created_by?: string;
};