// admin/src/api/generated/schemas/adminCreateAdminRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminCreateAdminRequest = {
  username?: string;
  name?: string;
  role?: string;
  password?: string;
  require_password_change?: boolean;
  two_factor_required?: boolean;
};